requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-classic>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.12.0
python-dotenv>=1.0.0
redis>=5.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
numpy>=1.24.0
networkx>=3.0
tiktoken>=0.5.0

